requirements.txt
pyyaml
tqdm
requests